Updated addFolderToStructure to include items as addItemsToStructure
