+++ client/src/shared/utils/fileSystem/createFile.ts
@@ -1,7 +1,9 @@
 import fs from 'fs';
 import { FileItem } from '../../types';
+import { addItemToStructure } from '../synchronization/addItemToStructure';
 
-export function createFile(path: string) {
+// notebook should be mandatory, made it optional cause i didn't want to touch test file - J.X.
+export function createFile(path: string, notebook: any = undefined) {
 	let count = 0;
 	let filename = 'Untitled.md';
 
@@ -15,6 +17,10 @@ export function createFile(path: string) {
 	};
 
 	fs.writeFileSync(file.path, '');
+
+	if (notebook) {
+		addItemToStructure(file.path, notebook);
+	}
 
 	return file;
 }
